# server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: adopt-animals-api
  description: Adopt Animals REST API
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
definitions:
  Animal:
    type: object
    title: Animal
    required:
      - type
      - gender
      - breed
      - birthdate
      - zone
      - peltColor
      - eyesColor
      - description
      - contact
    properties:
      id:
        type: integer
        example: 1
        readOnly: true
      name:
        type: string
        nullable: true
        example: 'Pepe'
      image:
        type: string
        example: 'assets/dog.jpg'
        readOnly: true
      type:
        type: string
        example: 'dog'
      gender:
        type: string
        example: 'male'
      breed:
        type: string
        example: 'Akita'
      birthdate:
        type: "string"
        format: "date"
      zone:
        type: string
        example: 'Buceo'
      peltColor:
        type: string
        example: 'Gris'
      eyesColor:
        type: string
        example: 'Negro'
      description:
        type: string
        example: 'Divertido, mimoso y bueno.'
      contact:
        $ref: "#/definitions/Contact"
  Contact:
    type: object
    title: Contact
    required:
      - name
      - email
      - phone
    properties:
      name:
        type: string
        example: 'John Doe'
      email:
        type: string
        example: 'john.doe@gmail.com'
      phone:
        type: string
        example: '+598 094135268'
paths:
...
  /animal/{id}:
    get:
      tags:
        - Animals
      description: Retrieves a specific animal by id
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the animal to retrieve
          type: integer
      responses:
        200:
          schema: 
            $ref: "#/definitions/Animal"
          description: Returns the animal with the specified id
        400:
          description: Invalid ID supplied
        404:
          description: Animal not found
    put:
      tags:
        - Animals
      description: Updates a specific animal
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the animal to retrieve
          type: integer
        - name: Animal
          in: body
          description: updated animal information
          required: false
          schema: 
            $ref: "#/definitions/Animal"
      responses:
        202:
          schema: 
            $ref: "#/definitions/Animal"
          description: Returns the updated animal
        400:
          description: Invalid input
        404:
          description: Animal not found
    delete:
      tags:
        - Animals
      description: Deletes an animal
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the animal to delete
          type: integer
      responses:
        204:
          description: Animal deleted
        400:
          description: Invalid ID supplied
        404:
          description: Animal not found
...